% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avrg.ratios.R
\name{avrg.ratios}
\alias{avrg.ratios}
\title{function to calculate average isotope ratios of fragments/ions of interest. I.e. isotope ratio divided by the number of relevant atoms in fragment/ion of interest}
\usage{
avrg.ratios(Iratios_results, ions, n.atoms)
}
\arguments{
\item{Iratios_results}{= data frame that results from applying the iratios() function}

\item{ions}{= vector which contains all fragments/ions of interest}

\item{atoms}{= vector wchih contains number of relevant atoms in fragment/ions of interest (ORDER MATTER!, has to be the same as in ions}
}
\value{
data.frame of isotope ratios divided by number of atoms of interest
}
\description{
function to calculate average isotope ratios of fragments/ions of interest. I.e. isotope ratio divided by the number of relevant atoms in fragment/ion of interest
}
