% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.shotnoise.R
\name{calc.shotnoise}
\alias{calc.shotnoise}
\title{Shot-noise limit calcualtion per scan. modified after Cajetan Neubauers script}
\usage{
calc.shotnoise(d.neff)
}
\arguments{
\item{d.neff}{= dataframe which is subset to to only contain one file of interest, ion of interest and one type of unsub/sub. count  and ratio calculation method to use for the SNL calculation. Doesnt work for Nio_b.avrg4_ratio and Nio_b.min20_ratio b/c rows contain NA. Have to remove them for SNL calc without affecting the other ratios}
}
\value{
table with shot noise limits (value) and number of effective ion counts (n.eff) per scan.
}
\description{
Shot-noise limit calcualtion per scan. modified after Cajetan Neubauers script
}
