% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cull.Idata.R
\name{cull.Idata}
\alias{cull.Idata}
\title{Function that filters the dataframe using different culling methods
currently culling on Nio values can be switched to Nio_weighted etc. etc.
1/2/3 standard deviations from mean ticxit
10%/20% MAD median(abs((ticxit)i-mean(ticxit)))}
\usage{
cull.Idata(filter.Idata_results, ratio, it)
}
\arguments{
\item{filter.Idata_results}{=results from filter.Idata function}

\item{ratio}{= isotope ratio calcualtion method preferred by user i.e. either "NL_ratio", "Nio_ratio", "Nio_w_ratio". Doesnt work for background subtracted ratios since negative values have to be replaced by NA first}

\item{it}{<= it value defined by user}
}
\value{
list containing dataframes that have been culled to include a subset of the input dataframe 1/2/3 standard deviations, 10%/20% MAD from the mean of Nio and below a certain it value. Output summary table can be used to evaluate different culling options for different fragments best hit for min. RSE
}
\description{
Function that filters the dataframe using different culling methods
currently culling on Nio values can be switched to Nio_weighted etc. etc.
1/2/3 standard deviations from mean ticxit
10%/20% MAD median(abs((ticxit)i-mean(ticxit)))
}
