% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.Idata.R
\name{filter.Idata}
\alias{filter.Idata}
\title{function that that will filter ilogs based on exact mass +- mass tolerance mmu, remove duplicate rows assigned to same ilog or rows that dont have all ilogs and add background level column per filename, compound, isotopolog}
\usage{
filter.Idata(import.Idata_results, tolerance)
}
\arguments{
\item{import.Idata_results}{result from import.Idata function}

\item{tolerance}{= user defined amount of deviation from exact mass (mass defect) permitted in daltons}
}
\value{
= dataframe only containing scans with mass defect as defined by tolereance value, unique assignment of one isotopolog per defined exact mass,
including only scan that have all defined isotopologs of interest per scan within peak elution time and background values for each filename,compound, ilog of interest (2 background filter methods)
}
\description{
function that that will filter ilogs based on exact mass +- mass tolerance mmu, remove duplicate rows assigned to same ilog or rows that dont have all ilogs and add background level column per filename, compound, isotopolog
}
