% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iratios.R
\name{I.ratios}
\alias{I.ratios}
\title{function that calautes isotope ratios from intensity values for each scan, also restructures the dataframe
ions.incremental is not = Eiler 2017 NIO calculation!!!
NL ratio, Nio ratio , Nio weighted ratio
ONLY WORKS FOR TWO ILOGS!}
\usage{
I.ratios(x)
}
\arguments{
\item{x}{results from filter.Idata function}
}
\value{
data frame contains substituted and unsubstituted iostopolog data in columns next to each other with isotope ratios (NOT DEVIDED by #Cs).Negative values after background subtraction are replaced with NA
}
\description{
function that calautes isotope ratios from intensity values for each scan, also restructures the dataframe
ions.incremental is not = Eiler 2017 NIO calculation!!!
NL ratio, Nio ratio , Nio weighted ratio
ONLY WORKS FOR TWO ILOGS!
}
