% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.Idata.R
\name{import.Idata}
\alias{import.Idata}
\title{function that allows to import data set extracted from raw input using isoX. Use combined.isox file for multiple acquisitions}
\usage{
import.Idata(df, ranges, masses, ions, ilogs)
}
\arguments{
\item{df}{=data frame of isox data table output}

\item{ranges}{= data frame of .txt file with start and end times for the peak of interest in each measurement run}

\item{masses}{= dataframe of isotopologs.tsv file which was also used to generate the .isox file}

\item{ions}{= vector which contains all fragments/ions of interest}

\item{ilogs}{= vector of isotopologs of interest (has to be contained in isotopologs.tsv file)}
}
\value{
= dataframe only containing columns for further analysis i.e. time ranges for each acquisition, filter for fragments/ions of interest and isotopologs of interest and mass defect value.
}
\description{
function that allows to import data set extracted from raw input using isoX. Use combined.isox file for multiple acquisitions
}
