% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.ratios.R
\name{s.ratios}
\alias{s.ratios}
\title{summary data for different NL ratio calculation options per compound
drops NA rows for background subtracted ratios
this function is very repetitive have to rewrite summarize with loops}
\usage{
s.ratios(avrg.ratios_results)
}
\arguments{
\item{avrg.ratios_results}{results of avrg.ratios function}
}
\value{
= summary data table for each ratio calculation type per compound and filename
}
\description{
summary data for different NL ratio calculation options per compound
drops NA rows for background subtracted ratios
this function is very repetitive have to rewrite summarize with loops
}
